When using web service, adding an appointment was causing an error.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3099 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/AccountingAutoPays.cs
@@ -35,7 +35,7 @@ namespace OpenDentBusiness{
 		///<summary></summary>
 		public static long Insert(AccountingAutoPay pay) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				pay.AccountingAutoPayNum=Meth.GetInt(MethodBase.GetCurrentMethod(),pay);
+				pay.AccountingAutoPayNum=Meth.GetLong(MethodBase.GetCurrentMethod(),pay);
 				return pay.AccountingAutoPayNum;
 			}
 			if(PrefC.RandomKeys) {
