+++ OpenDentBusiness/Data Interface/AppointmentRules.cs
@@ -34,7 +34,7 @@ namespace OpenDentBusiness{
 		///<summary></summary>
 		public static long Insert(AppointmentRule rule) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				rule.AppointmentRuleNum=Meth.GetInt(MethodBase.GetCurrentMethod(),rule);
+				rule.AppointmentRuleNum=Meth.GetLong(MethodBase.GetCurrentMethod(),rule);
 				return rule.AppointmentRuleNum;
 			}
 			if(PrefC.RandomKeys) {
