+++ OpenDentBusiness/Data Interface/ApptViewItems.cs
@@ -38,7 +38,7 @@ namespace OpenDentBusiness{
 		///<summary></summary>
 		public static long Insert(ApptViewItem Cur){
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				Cur.ApptViewItemNum=Meth.GetInt(MethodBase.GetCurrentMethod(),Cur);
+				Cur.ApptViewItemNum=Meth.GetLong(MethodBase.GetCurrentMethod(),Cur);
 				return Cur.ApptViewItemNum;
 			}
 			if(PrefC.RandomKeys) {
