+++ OpenDentBusiness/Data Interface/CanadianExtracts.cs
@@ -65,7 +65,7 @@ namespace OpenDentBusiness{
 		///<summary></summary>
 		private static long Insert(CanadianExtract cur) {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				cur.CanadianExtractNum=Meth.GetInt(MethodBase.GetCurrentMethod(),cur);
+				cur.CanadianExtractNum=Meth.GetLong(MethodBase.GetCurrentMethod(),cur);
 				return cur.CanadianExtractNum;
 			}
 			if(PrefC.RandomKeys) {
